+++ CHANGES.rst
@@ -4,6 +4,9 @@
 - Break potential reference cycle between ``request`` and ``context``.
   See https://github.com/Pylons/pyramid/pull/3649
 
+- Remove ``update_wrapper`` from ``pyramid.decorator.reify``.
+  See https://github.com/Pylons/pyramid/pull/3657
+
 2.0b0 (2020-12-15)
 ==================